+++ src/YOUR_NAME_ex/ex03/code/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+__author__ = "Jørgen Kongsro"
+__email__ = "jorjoh@nmbu.no"
+__version__ = "0.0.1"
+
+from .median import median
+from .bubble_sort import bubble_sort
